Break up variables.go file into smaller files
